reduce further to 16 bytes?

+++ crates/wasmparser/src/validator/component.rs
@@ -1517,7 +1517,7 @@ impl ComponentState {
         let current = components.last_mut().unwrap();
         check_max(current.type_count(), 1, MAX_WASM_TYPES, "types", offset)?;
 
-        current.core_types.push(TypeId::new(ty.type_size, ty.index, Some(current.core_types.len()), true));
+        current.core_types.push(TypeId::new(ty.type_size as usize, ty.index, Some(current.core_types.len()), true));
 
         Ok(())
     }
@@ -1529,7 +1529,7 @@ impl ComponentState {
         let current = components.last_mut().unwrap();
         check_max(current.type_count(), 1, MAX_WASM_TYPES, "types", offset)?;
 
-        current.types.push(TypeId::new(ty.type_size, ty.index, Some(current.types.len()), false));
+        current.types.push(TypeId::new(ty.type_size as usize, ty.index, Some(current.types.len()), false));
 
         Ok(())
     }
